+++ components/SearchScreen/NoResultsOrError/index.tsx
