+++ components/SearchScreen/SearchFooter/index.tsx
